requirements.txt
mss
pytesseract
opencv-python
Pillow
numpy
requests
python-telegram-bot